Fixed INTERNAL-mode connection bug in dlt-daemon
When logging-mode of the DLT-daemon was set to DLT_USER_MODE_INTERNAL or DLT_USER_MODE_OFF and a client tried to connect to the daemon,
the socket for accepting new connections would be closed. This would make connecting to the daemon impossible, even if the logging mode
was reset to DLT_USER_MODE_EXTERNAL or DLT_USER_MODE_BOTH.

Fixed this by removing conditional check on the logging mode - clients connections are always accepted now. If the logging-mode is
being switched from INTERNAL to EXTERNAL or BOTH while a client is connected, the client will then start to receive messages normally.

Signed-off-by: Alexander Wenzel <Alexander.AW.Wenzel@bmw.de>