DltLogstorage: Logstorage Cache
When using DltLogstorage on internal storage device, it is needed to
reduce writing to internal storage device as much as possible.

This patch introduces sync strategies to Logstorage to provide that
functionality.
The ON_MSG strategy is the default sync strategy that flushes every
written log message to the storage device (fflush).
The ON_DAEMON_EXIT strategy only flushes data to disk when the daemon
exits.

The strategy can be defined per filter in the dlt_logstorage.conf
configuration file by adding
SyncBehavior=<Strategy> to a configuration.

Signed-off-by: Christoph Lipka <clipka@jp.adit-jv.com>