MultiNode: Send control messages after connection
A list of control messages can be specified in the dlt_gateway.conf.
These messages then are send immediately to the passive node after the
connection has been established.

Currently supported control messages are
- GET_LOG_INFO
- GET_SOFTWARE_VERSION

Having the "SendControl" property is now mandatory for a passive node
configuration, but the list of control messages might be empty.

Signed-off-by: Christoph Lipka <clipka@jp.adit-jv.com>

Change-Id: I0ae4501449bf264c2d6f32724b0f035926115611