Bugfix: dlt-system filetransfer crash
If a file get written to after it appeared in a folder that is watched
by the dlt-system filetransfer thread, the application tries to send
this file multiple times, causing it to crash. This patch fixes this by
canceling the sending process more gracefully if the file cannot be
found.

Change-Id: Id19b72061bed41642d12cd33afeaa41fb444b2c8
Signed-off-by: Lutz Helwing <lutz_helwing@mentor.com>