dlt-system filetransfer waits for a client
For each file that is supposed to be sent, dlt-system will check if a
client is connected to the daemon. If not, it waits and keeps checking
once per second, until a client is connected.

Files will still be compressed and moved to the .tosend subdirectory,
regardless of the client connection.

Changed the default startup timeout in the configuration file to 0.

Change-Id: I5850adb51c381bc6279ac59f5bd985491ba5caf3
Signed-off-by: Lutz Helwing <lutz_helwing@mentor.com>