Revert truncation of string or raw block
Commit d319ebf4d2d449d70befae0eebbdd82c19c5315f introduced new handling
of the situation when writing data to the user buffer would exceed its
size. This was achieved by truncating the data to fit. This solution had
a problem:

It could lead to a buffer overflow when the updated arg_size gets negative
(actually undefined value because it is unsigned).

As a quick fix the behaviour is reverted to the state before the
aforementioned commit.

Signed-off-by: Lutz Helwing <lutz_helwing@mentor.com>