process user message: Fix bound handling
On context and application registration, the data needs to be read in 2
parts. If the first part was at the end of the receiver buffer, it was
wrongly removed whereas the complete data is not full received and
therefore interpretable. This has to be fixed by not removing the data
before everything is checked.

Thus, while readjusting the buffer to the first found header, the offset
was not properly updated.

Signed-off-by: Frederic Berat <fberat@de.adit-jv.com>

Process User Messages: Fix buffer handling

In case application and context registration the buffer need to be read
two times. But before removing the data, the data need to be readjusted
to its origin. This was missing in a previous fix.

Furthermore, in case of corrupted description field, applications and
contexts will be registered anyway with the dummy description "Unknown".

Signed-off-by: Christoph Lipka <clipka@jp.adit-jv.com>