dlt-daemon: Output signal number at exit
So far, signal number had been output in signal handler.
However this was not so safety. And when syslog was used as log
output destination of dlt_log(), we could not see any info in the
same Life Cycle.
So dlt_log() is removed from signal handler and signal number is
newly output by DLT message and dlt_log() at exit.
This info will help analysis why dlt-daemon exits.

Signed-off-by: Yusuke Sato <yusuke-sato@apn.alpine.co.jp>