DLT_FORK_Child_LOCKUP
As current implementation takes care of not Initializing the new set of buffers for the forked process, but the forked process would still contains the copy of shared resources and memory, which were not freed during the exit of child process. Calling dlt_free from child process would result in dead lock situation if parent process has locked on the resource during the child creation. So, to overcome the dead lock situation, allocating the memory for mutex at run time instead being the static and freeing it from the child process and deleting it only from the parent process and added a new function dlt_force_free to free the resources from the child process which doesn't have any locking mechanism.