Pipe name to use APID instead of PID
Currently DLT library uses PID of the process as a part
of the pipe's name which is created in dltpipes folder.

This should be changed to support PID namespace which means
that PID of the process can't be part of the name. We propose to
use DLT APP ID as a replacement because that is always unique per
user (client).