Updated Made writev reliable for fixing data corruption
On random scenario's  "writev()" writes only part of the DLT data to Daemon FIFO due to un-availability of space in Daemon FIFO,  and returns PIPE Error which in turn stores the same complete trace in the applications ring buffer.
Incomplete trace was read by Daemon(from FIFO) and routed the same to DLT Clients which in turn results in Data corruption at DLT Viewer due to non-availability of complete trace as per the DLT Payload headers.
Hence proposed solution would check for space availability for each trace prior writing the data to Daemon FIFO, during the failure conditions the data to be stored internally in its application ring buffers and ensuring incomplete traces were not written on to Daemon FIFO.