daemon: Loop for client fds
In dlt_daemon_client_send_all_multiple(), while in the loop,
the connection to client may be destroyed by recursive call to
dlt_daemon_close_socket(). So in some scenario we still use the removed
connection, which is unexpected behavior.

Solution: loop on the fds array and find the client connection
corresponds to.

Signed-off-by: Vo Trung Chi <Chi.VoTrung@vn.bosch.com>