Do not install systemd service files for binaries that are not built (#129)
* Do not install systemd service files for binaries that are not built

Noticed that dlt-system.service unconditionally failed since the binary
was missing when building with a default configuration after an upgrade
from 2.17.0 to 2.18.1. WITH_DLT_SYSTEM default changed to OFF in
1b3c8b3d10fbc9a8dd9c717a939babd901b0b243 .

Fix by only installing dlt-system.service if dlt-system is built and do
the same for other binaries that have a service file. Also install() is
moved to if() blocks used for configure_file().

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>