dlt-daemon: Added functionality to use specific set of bindAddresses (IP addresses) for which the dlt-daemon will accept connections.
I added the possibility to have in dlt.conf an option that specifies a set of IP addresses.
The daemon will do a socket bind only for this list, therefore external connections will be limited to this set of IP addresses.
If this option is not given, the default INADDR_ANY is used ("0.0.0.0") as before.

The option in dlt.conf can look like this:
BindAddress = 160.48.199.98;160.48.199.97;160.48.199.226,160.48.199.186;160.48.199.139;172.16.222.99

There can be as many addresses but only specified on one single line. They have to be delimited by ',' or ';'. ' ' (space) delimiter does not work.
IPv6 addresses are supported as well. For instance:

BindAddress = fe80::255:7bff:feb5:7df7

There can be IPv4 addresses and IPv6 in the same time, but if at least one IPv6 address are used, dlt-daemon must be compiled with DLT_USE_IPv6 flag ON.