libdlt: make the register-application work only once
I am developing a middleware with a single library. And I am using the dlt to output the logs.
But I find some of the executables that using my library was not using dlt, it means none will register the application, and my library could not output any logs.
But if I register the application in my library, my application id will override the application id of the executable who using dlt.

So, I did this change to make the register-application only work once.
If the excutable registered the application, my library 's registering will be ignored.
Only thing might change is, the executable need to register the application as soon as possible, early before any libraries.

Signed-off-by: Charles Chan <charles@zeerd.com>