libdlt: add an interface to get the status of application registering
I am developing a middleware with a single library. And I am using the dlt to output the logs.
But I find some of the executables that using my library was not using dlt, it means none will register the application, and my library could not output any logs.
But if I register the application in my library, my application id will override the application id of the executable who using dlt.

I added this new interface to get the status of application registering.
So that I could decide if I need to register the application in my library or not.

Signed-off-by: Charles Chan <charles@zeerd.com>