Change the DLT_CHECK_RCV_DATA_SIZE macro to an internal function
The use of this macro in a if condition is actually non-standard C++
code. GCC issues a warning that ISO C++ forbids braced-groups within
expressions when compiled with -Wpedantic. Other compilers fail with an
error. Change the macro to an internal function in order to improve the
portability to other platforms (i.e. MSVC, proprietary embedded).

Signed-off-by: Sebastian Lipponer <mail@sebastianlipponer.de>