daemon: logstorage with wildcards
In the current implementation, if FILTER section is setting
wildcard for both "LogAppName" and "ContextName", it will be ignored.
But it is still a valid configuration for NON-VERBOSE.

In order to consolidate the daemon's behavior, the combination of wildcard
for both "LogAppName" and "ContextName" in FILTER section will be treated
as only ECUID is specified.

Signed-off-by: Bui Nguyen Quoc Thanh <Thanh.BuiNguyenQuoc@vn.bosch.com>