gateway: Improvement of handling Gateway config
In the current implementation, we believe in user to treat
General section as Mandatory. Hence system could be reset
due to memory of connections incorrectly.

At the moment, General could be handled as Optional with default value
of interval. It means General could be absent from config file.

Improvement is to raise WARNING to user, set default value and
update the number of connections properly.

Signed-off-by: Bui Nguyen Quoc Thanh <Thanh.BuiNguyenQuoc@vn.bosch.com>