offline storage: Improvement log messages at bottom
- In case there are several filters in config dlt_logstorage.conf
which are sharing the same file name, the dlt-daemon could not handle
to write the dlt messages at the proper file. It means the latest dlt
messages might be not written into the newest file.

- In order to resolve this, dlt-daemon must keep the newest file for
each filename. Everytime new dlt message is going to be written,
dlt-daemon must ensure that it must be written out to the newest file.

Signed-off-by: Bui Nguyen Quoc Thanh <thanh.buinguyenquoc@vn.bosch.com>
Signed-off-by: KHANH LUONG HONG DUY <khanh.luonghongduy@vn.bosch.com>