dlt-daemon: log levels are not controlled
The issue happens in case:
dlt_logstorage.conf is set available at startup via dlt.conf
and both LogAppName and ContextName are wildcard.

In such a case, when logstorage disconnection happens,
dlt-daemon is missing a check for case in which both are
wildcard.

It causes the reset loglevel is not happening.
Then the control message to change loglevel will be
handled incorrectly.

Signed-off-by: Bui Nguyen Quoc Thanh <thanh.buinguyenquoc@vn.bosch.com>