Avoid memory access errors with 4-chars context ids
For a 4-chars CTXID, i.e. one that does not include a null character,
the strlen() calls were happily accessing memory until they eventually
encountered a null character somewhere in memory.

This was detected by valgrind, which reported a memory error when
using a CTXID such as "INTM":

==21924== Conditional jump or move depends on uninitialised value(s)
==21924==    at 0x4C30F78: strlen (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==21924==    by 0x4E4B5A0: dlt_print_id (dlt_common.c:303)
==21924==    by 0x4E4CF47: dlt_message_header_flags (dlt_common.c:687)
==21924==    by 0x4E51434: dlt_message_print_ascii (dlt_common.c:3169)
==21924==    by 0x4E4247A: dlt_user_print_msg (dlt_user.c:4108)
==21924==    by 0x4E46D92: dlt_user_log_send_log (dlt_user.c:3670)
==21924==    by 0x4E46F14: dlt_user_log_write_finish (dlt_user.c:1611)

Sanitize some code

Using memset() and memcpy() is always preferable to hand-rolled loops,
because compilers have built-in support for them.

Signed-off-by: Martin Willers <M.Willers@gmx.net>