dltuser: add method to configure ring buffer size per application.
Logs of applications are kept in a ring buffer.
Each application (ie user) using logging functionality
of dlt-daemon should be allowed to write
to ring buffer a limited asize of logs not to exceed
limits available in terms of memory and performance.
This patch is useful for upstream.
Because, first, the ability of limiting logs
of each application will make the use of system memory
resources without exceeding the overall capacity limits.
Secondly the logs of each pplication will be limited
as a guard to excessive logging - ie logging in
for loops- which could not easily
be determined since most of them occur in extreme conditions.

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>