dlt-receive: flush stdout buffer by signal
When dlt-receive logs is redirecting to file then a terminating signal
is sent to dlt-receive, it will stop without flushing all logs from
stdout to file, due to the default size of stdout is 4096 byte.

When dlt-receive receive the SIGTERM or SIGINT, it flushes stdout buffer
to stdout before terminating.

Signed-off-by: KHANH LUONG HONG DUY <khanh.luonghongduy@vn.bosch.com>