Re-check the ring buffer size after incresing it
If the buffers are bigger than the free space in the ringbuffer, the
ringbuffer is increased by a step size set at initialization.
But there is no check if this increase was sufficient.
Fix this by using a while loop that increases the ring buffer size until
it is big enough or the buffer can not be further increased.

Signed-off-by: Jan Schrewe <schrewe@uni-bonn.de>