dlt_user: fix invalid poll timeout
The poll timeout was only set for fifo.
TCP connections require this timeout as well.
dlt_user_log_check_user_message also validated the file descriptor
to be greater 0.
Because 0 is a valid file descriptor this check has been changed
to greater or equal 0.
poll receives a timeout in milliseconds.
The given paramter was in nanoseconds.
An additional define takes adds the delay in miliseconds.

Signed-off-by: Alexander Mohr <alexander.m.mohr@daimler.com>