bug-fix: fix invalid file descriptor check
Some functions validated file descriptor to be greater 0.
If a process is started without stdin, stdout and stderr
the first file descriptor allocated by the process will be 0.
This also will be the case if the above mentioned file descriptors
will be closed on purpose.
As 0 is a valid fd, some methods had to be changed to reflect this.

Signed-off-by: Alexander Mohr <alexander.m.mohr@daimler.com>