dlt_client_main_loop running in an infinite loop restricts graceful exit of DLT Client code
Issue:
------
dlt_client_main_loop currently uses an infinite loop ("while (1)"). This
creates problems when a DLT client is running in a thread as part of a larger
application. Graceful exit (for example: during object destruction) is not
possible because a thread is running dlt_client_main_loop in the background.

It is also not possible to exit the client, if we want it to fetch only a
pre-decided number of messages.

Solution:
---------
Allow user to define a callback function to check whether the next message
should be fetched.

dlt-test-client.c has a new option to fetch a pre-decided number of messages,
after which the client will exit.

Signed-off-by: Sreeharsha Ramanavarapu <sreeharsha.r@gmail.com>