Better formatting of RAWD
Previously, the bytes of a RAWD argument were be printed onto the console
as space-separated hex values, making it difficult or impossible to see
where the RAWD argument ends and the next argument starts. For instance,
with code like this:

    dlt_user_log_write_raw(&ctxData, "data", 4);
    dlt_user_log_write_uint8(&ctxData, 53);

the console output is shown as:

    [64 61 54 61 53]

and there is no indication that there are two arguments, one RAWD and
one UINT8.

This patch changes for console formatting of RAW data to separate by
apostrophs, making the above message appear as:

    [64'61'54'61 53]

This makes it immediately obvious that there is a RAWD argument, and
still keeps the separators quite unobtrusive.

Signed-off-by: Martin Willers <M.Willers@gmx.net>