daemon: Improve signal handler for timers
In the initial implementation of timer_thread, the signals
SIGUSR1 is blocked and only checked via sigpending.
There will be an issue if the interval of timers are quite long.
It could cause the main thread waits for a long time in case of
termination gratefully.

Fix: let the timer_thread be interrupted by signal SIGUSR1 and
handle it properly then.

Signed-off-by: Bui Nguyen Quoc Thanh <thanh.buinguyenquoc@vn.bosch.com>