dlt_user: Make dlt_init thread safe
This commit introduces the use of atomic_compare_exchange at beginning
of both dlt_init functions. This function checks in an atomic way, if
'dlt_user_initialised' is 'false' (first run through dlt_init) and directly
sets it 'true'. If a second thread also entered dlt-init at this point of time,
its call of atomic_compare_exchange will return false and dlt_init will be
immediately left due to that. This is only interesting for startup, because
before each call of dlt_init, there is a check on 'dlt_user_initialised' that
will fail for all later executions.

Signed-off-by: Darian Biastoch <dbiastoch@de.adit-jv.com>