fix -Wformat issues reported by clang
* some previous dlt_vlog was using %lu specifier to display an uint64_t
variable. However, this isn't portable as depending on the system you're
targeting (32bits, 64bits, Linux, android, etc.), the uint64_t can be an
unsigned long long. This prevents compilation on system where -Werror
and -Wformat is activated coupled with a less tolerant compiler
(e.g clang). Instead, PRIxN macros (e.g PRIu64 for uint64_t) are now
used as specifier to ensure compatibility for each platforms
* for bytes_read which is of type ssize_t, zd specifier is used as
indicated in C99 standard
* for st_size (from struct stat) which is of type off_t, POSIX says that
it should be a signed integer. To be safe, we now use %jd specifier to
ensure that it will fit correctly