dlt-daemon: Only create directories if they do not exist yet
dlt_mkdir_recursive runs mkdir for directories which already exist.
This fails if the directory already exists and the dlt-daemon
does not have the permissions to change the mode on the directory.
On many systems the directories are not created by dlt-daemon
itself but rather through the build system or while the
root filesystem is created.
At this stage the permissions are already set properly
and they might be less permissive than what dlt-daemon expects
but still sufficient.
For example is rwx not necessary for each directory in the tree.

Before this commit the following scenario would fail:
* Path is set to /mnt/dlt/
* mnt has 755 permissions for root
* dlt has 755 for dlt user

As dlt daemon would try to run mkdir on /mnt
dlt_mkdir_recursive would fail due to lack of permissions
although the permissions are sufficient for dlt-daemon to
work and write its files into /mnt/dlt/
With this commit /mnt would not be created anymore because it
already exists.
If the permissions are not sufficient for dlt-daemon to write in
it will fail when trying to create files and exit at this point

Signed-off-by: Alexander Mohr <alexander.m.mohr@daimler.com>