dlt-shared: add dlt_log_user with timeout
if a client is not properly processing data
i.e. due to the client itself being slow or a
slow network connection, data should be
dropped instead of slowing the dlt-daemon down.
The later could lead to a systemd watchdog timeout
or lost logs because the input buffer of the daemon
will be filled completely.

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>