Add Journal Gateway Mode
Adds a new connection type which reads the Systemd Journal log stream
for current boot and forwards everything to connected clients. This new
mode is useful on Linux system that predominantly use the Journal for
logging, which dlt-daemon cannot replace outright. It's enabled by
adding 'JournalGatewayMode = 1' to dlt.conf.

This mode of dlt-daemon is similar to dlt-system's Journal Adapter,
enabled by 'JournalEnable = 1' in dlt-system.conf. However, this new
implementation preserves certain Journal meta-data in DLT formatting
to facilitate filtering via control messages:

* App ID is the first 4 alpha-numeric chars of SYSLOG_IDENTIFIER.
  Examples:
    "httpd" ==> "HTTP"
    "wpa_supplicant" ==> "WPAS"
    "ls" ==> "LS<NULL><NULL>"

* Context ID is translated from SYSLOG_FACILITY enum.
  Examples:
    LOG_KERN ==> "KERN"
    LOG_DAEMON ==> "DAEM"
    LOG_FTP ==> "FTP<NULL>"

* Level is translated from PRIORITY enum as follows:
    LOG_EMERG, LOG_ALERT, LOG_CRIT ==> DLT_LOG_FATAL
    LOG_ERR ==> DLT_LOG_ERROR
    LOG_WARNING ==> DLT_LOG_WARN
    LOG_NOTICE, LOG_INFO ==> DLT_LOG_INFO
    LOG_DEBUG ==> DLT_LOG_DEBUG

Level filtering in this mode is implemented by the Journal read loop.
Each log entry is checked against a hash map of App ID to limit or the
default if not present in map. The hash map and default can be updated
by control messages from clients.

Signed-off-by: Haris Okanovic <harisokn@amazon.com>