Limit set of symbols exported from libdlt
Summary:

This commit deals with issues potentially introduced in code
linked to shared libdlt. The library previously exported all
symbols by default, up to ~60 of which were not prefixed 'dlt*'
and could conflict with symbols in the consumer code.

After applying this commit only symbols prefixed 'dlt*'
can be observed using

$ nm --dynamic --defined-only ./build/src/lib/libdlt.so | cut -f3 -d' ' | sort

which drastically reduces the chance of such conflicts.

Brief:

- explicit symbol export using DLT_EXPORT macro as provided by
  CMake's GenerateExportHeader module
- rename get_filename_ext to dlt_get_filename_ext
- rename getFileSerialNumber to dlt_get_file_serial_number
- hide dlt_daemon's original main() in unit tests to avoid
  multiple definition on some platforms
- add CMake 3.12's GenerateExportHeader to repository for
  backward-compatibility with pre-3.12 CMake (i.e. Ubuntu 18.04)