Fix DLT-Multinode: Gateway does not recognize reset of passive node #551
Enable TCP keepalive to detect broken connections due to network disconnects.
Without this in cases where no TCP FIN or RST package is received from the server, the dlt_client will never notice that it was disconnected from the server. Thus it will not reconnect when the server is available again.
Also improve logging output to see which connection fails.