Close file after write
This commit is to avoid keeping the log file open when not needed.
That is helpful if it is a gz file as it then can be copied without
being truncated. If the file is copied at the very same time as it
is being written to it will still be truncated, but with this commit
that risk is smaller.