Fix compiler warnings related to strings
Fix string truncation and overflow warnings so the compiler flags
stringop-overflow and stringop-truncation can be enabled.

This shouldn't change any behaviour nor fix any bugs. It is just
to be able to run the compiler with those flags so future bugs
can be avoided. However I have added some extra checks for string
lengths.

Most changes is to avoid using strncat and instead use strcat. This
doesn't make the code safer but removes the false safeness of using
strncat wrongly.