watchdog: improve the behavior how watchdog is handled
* dlt_daemon_process_user_messages
* dlt_daemon_send_ringbuffer_to_client
* dlt_daemon_client_send_all_multiple
* dlt_daemon_process_systemd_timer
* dlt_daemon_handle_event
* dlt_daemon_socket_sendreliable

These methods can take longer than the systemd watchdog timeout allowed, depending on the number of clients, which leads that the systemd watchdog kills dlt-daemon.
Add a timeout to these methods so that they may only run for as long as the watchdog timer is set.
In addition, set the default WatchdogTimeout to an appropriate value in case the env variable is not set.

Signed-off-by: Ahmet Findikci <ahmet.findikci@mercedes-benz.com>