Revert "dlt-daemon: Handle partial message parsing in receiver buffer"
This reverts commit 1737de27825c691e925a7d42abe34311159c48c6.

The original commit aimed to address potential partial message
reception. However, this issue is not observed on Linux and QNX systems,
which are the primary target platforms for dlt-daemon.
Furthermore, the changes introduced by the reverted commit have been
found to cause message loss in certain scenarios.
Therefore, to ensure the stability and reliability of dlt-daemon, this
commit reverts the problematic changes.

This revert addresses issue #631.